+++ src/textual/widget.py
@@ -3290,12 +3290,14 @@ class Widget(DOMNode):
     def _on_focus(self, event: events.Focus) -> None:
         self.has_focus = True
         self.refresh()
-        self.post_message(events.DescendantFocus())
+        if self.parent is not None:
+            self.parent.post_message(events.DescendantFocus(self))
 
     def _on_blur(self, event: events.Blur) -> None:
         self.has_focus = False
         self.refresh()
-        self.post_message(events.DescendantBlur())
+        if self.parent is not None:
+            self.parent.post_message(events.DescendantBlur(self))
 
     def _on_mouse_scroll_down(self, event: events.MouseScrollDown) -> None:
         if event.ctrl or event.shift: